cppcheck Bundle 59: tasks/+network/ uselessAssign + dead-clause sweep
Triage from .audit_cache/cppcheck-b59-{network,tasks,fixes}.xml — 13
actionable sites + 2 FP classes.  Prior bundles already closed every OOM
NULL-check, %d/%u format-string, and deref-before-check site in
network/cloud_sync/ and tasks/, so post-fixes scan is at zero
error/warning over the changed files.

Useless assignment to function parameter (9 sites — same `param = NULL;`
after `free(param)` shape as Bundle 58's `core_info.c:2802` fix; C passes
parameters by value so the assignment is dead and cppcheck rightly flags
it):

  - tasks/task_autodetect.c:93                 autoconfig_handle = NULL;
  - tasks/task_core_backup.c:144               backup_handle = NULL;
  - tasks/task_core_updater.c:268              list_handle = NULL;
  - tasks/task_core_updater.c:682              download_handle = NULL;
  - tasks/task_core_updater.c:1180             update_installed_handle = NULL;
  - tasks/task_core_updater.c:1607             pfd_install_handle = NULL;
  - tasks/task_core_updater.c:1941             pfd_switch_cores_handle = NULL;
  - tasks/task_database.c:1762                 manual_scan = NULL;
  - tasks/task_menu_explore.c:66               menu_explore = NULL;
  - tasks/task_pl_thumbnail_download.c:405     pl_thumb = NULL;
  - tasks/task_playlist_manager.c:89           pl_manager = NULL;

All eleven are the last statement of a `free_X(X *param)` helper whose
sole caller is the task-state free-callback — the assignment touches a
function-local copy of the pointer that goes out of scope at the closing
brace.  Dropped the dead line in each.

Dead-clause / redundant-condition cleanups (3 sites):

  - tasks/task_content.c:573                   `if (archive_delim && *archive_delim)`
                                               archive_delim was already proven
                                               non-NULL at the surrounding
                                               `if ((archive_delim = path_get_archive_delim(path)))`
                                               at line 555, and `archive_delim++`
                                               at line 572 cannot null a non-NULL
                                               pointer.  Dropped the redundant
                                               first conjunct — left `if (*archive_delim)`.
  - tasks/task_pl_thumbnail_download.c:830     `if (pl_thumb)` guard inside the
                                               `case PL_THUMB_ITERATE_TYPE` arm,
                                               immediately after the same arm
                                               already dereferenced pl_thumb at
                                               line 825 in
                                               `task_set_progress(task, ((pl_thumb->type_idx - 1) * 100) / 3);`.
                                               pl_thumb is proven non-NULL at
                                               function entry (line 757-758:
                                               `if (!(pl_thumb = ...)) goto task_finished;`).
                                               Stripped the dead guard and the
                                               now-stale `/* fold into the
                                               guarded block */` comment it was
                                               trying to justify.
  - tasks/task_translation.c:548               `scaler->in_stride = -1 * width * 3;`
                                               where `width` is `unsigned` and
                                               `in_stride` is `int`.  Without
                                               the cast `-1 * width` is
                                               unsigned-promoted to UINT_MAX
                                               and the negative-stride encoding
                                               (used to flip the image
                                               vertically for the scaler)
                                               relies on implementation-defined
                                               unsigned-to-signed truncation.
                                               Cast width to (int) so the
                                               multiplication happens in signed
                                               space — `-1 * (int)width * 3`.

False-positive suppressions documented at-line in
`.cppcheck-suppress.txt` (S16, S17 — the rule that each entry must be
anchored to a line so a future rewrite forces re-triage stays):

  - S16 preprocessorErrorDirective:network/net_logger.c:33,37
    `net_logger.c` requires PC_DEVELOPMENT_IP_ADDRESS / _UDP_PORT to be
    defined in CFLAGS at build time.  Without them the netlogger driver
    cannot dial the developer's logging host, so the `#error` is a
    deliberate hard-fail to prevent a silent no-op build.  The .c file
    is only compiled when qb's --enable-netlogger picks it up, so the
    `#error` is a precondition gate, not a defect.

  - S17 uninitvar:network/netplay/netplay_frontend.c:2305
    `netplay_input_state_for()` takes `netplay_input_state_t *list`
    where `netplay_input_state_t` is itself a pointer type, so `*list`
    is a pointer the function walks as a linked-list head.  Every call
    site passes `&delta->resolved_input[i]` or similar — the head field
    on a delta struct that was zero-initialised at allocation, so
    `*list` is always defined (NULL or a node).  cppcheck doesn't infer
    the caller-side initialisation.

Verified via:

  $ make -j4    # build clean (only pre-existing -Wstringop warnings in
                #  slang_cache.cpp and video_shader_parse.c, unrelated)
  $ cppcheck --enable=... -j 4 \
        tasks/task_autodetect.c tasks/task_content.c \
        tasks/task_core_backup.c tasks/task_core_updater.c \
        tasks/task_database.c tasks/task_menu_explore.c \
        tasks/task_pl_thumbnail_download.c tasks/task_playlist_manager.c \
        tasks/task_translation.c \
        network/net_logger.c network/netplay/netplay_frontend.c
  0 error/warning  (was: 14 error/warning before fix + 2 FP)

Aggregate from initial network/+tasks/ scan: prior bundles (52, 56, 57,
the OOM-NULL-check 794a7a04bf, the atomic-write b510d9d654) had already
closed 47 of the 61 raw findings; this bundle closes the remaining 14
actionable and 2 FP classes.

Refs:
  audit branch: .audit_cache/cppcheck-b59-{network,tasks,fixes,postfix}.xml

+++ tasks/task_translation.c
@@ -545,7 +545,7 @@ static void handle_translation_response(
          scaler->out_height    = height;
          scaler->scaler_type   = SCALER_TYPE_POINT;
          scaler_ctx_gen_filter(scaler);
-         scaler->in_stride     = -1 * width * 3;
+         scaler->in_stride     = -1 * (int)width * 3;
 
          scaler_ctx_scale_direct(scaler, raw_output_data,
                (uint8_t*)raw_image_data + (image_height - 1) * width * 3);
